FIX: The fixed maintenance parameter for vehicles would behave erratically when numbers > 255 were used. (This is a fix to Makeobj-Experimental and Simutrans-Experimental) FIX: Crashes in some cases when setting up a new line. FIX: Compile errors on 64-bit platforms

+++ besch/vehikel_besch.h
@@ -72,7 +72,7 @@ private:
 	uint16 gewicht; //Weight
 	uint32 leistung; //Power
 	uint16 betriebskosten;  //Running costs
-	uint16	fixed_maintenance; //@author: jamespetts, April 2009
+	uint32 fixed_maintenance; //@author: jamespetts, April 2009
 
 	uint16 intro_date; // introduction date
 	uint16 obsolete_date; //phase out at
